ㄴㅁㅊ

+++ D3Dx/cCollision.h
@@ -36,9 +36,9 @@ public:
 	bool PlayerPlayerCollision(D3DXVECTOR3& pos1, D3DXVECTOR3& pos2, D3DXVECTOR3& dir1, D3DXVECTOR3& dir2);
 	bool PlayerObjectCollision(list<cIGObj*>::iterator iter, D3DXVECTOR3& pos1, D3DXVECTOR3& pos2, D3DXVECTOR3& dir1, D3DXVECTOR3& dir2);
 	bool ObjectObjectCollision(list<cIGObj*>::iterator iter1, list<cIGObj*>::iterator iter2);
-	bool WallCollisionX(int moveX, D3DXVECTOR3& pos, D3DXVECTOR3& dir);
-	bool WallCollisionZ(int moveZ, D3DXVECTOR3& pos, D3DXVECTOR3& dir);
-	bool WallVertexCollision(int moveX, int moveZ, D3DXVECTOR3& pos, D3DXVECTOR3& dir);
+	bool WallCollisionX(float moveX, D3DXVECTOR3& pos, D3DXVECTOR3& dir);
+	bool WallCollisionZ(float moveZ, D3DXVECTOR3& pos, D3DXVECTOR3& dir);
+	bool WallVertexCollision(float moveX, float moveZ, D3DXVECTOR3& pos, D3DXVECTOR3& dir);
 	cIGObj* DetectObject(int playerNum);
 	bool DetectMovement(int& moveX, int& moveZ, D3DXVECTOR3 dir);
 	float GetDistance(D3DXVECTOR3 vec1, D3DXVECTOR3 vec2);
